Move monolith bundler to its own file
